Added layers of abstraction for contexts. Added levels. Added very beginnings of level editor
Needs some work with the snap to grid ("nearest" function in "terrain.lua")

Click to switch to editor when in game. Spacebar to switch back to game.

Note that there is a lot to be fixed at this point, but I need to go to bed.
Terrain does not have collision, I wanted it to look good and work well from a creation standpoint,
before implementing collision with edgeBody objects (collision enabled line segments).

In the future, we can subdivide them, made nice curves, and so on. For now, it's pretty ugly.
Needed a place to start.