Here's the original main.lua I built in a testbed for the terrain and snap-to-grid
I would ask that people look at this before changing the other terrain, because this shows how it is SUPPOSED to work.

Obviously, it's a little broken in the main game. If you think the core functionality as it was intended should change, go for it, but at least see how it's supposed to work before removing the current code ("It doesn't work so I'll remove it")