Terrain creation and editing now works. No collision/smoothing yet
Camera has a zoom! mouse scroll in editor to try it out.
I may change camera back to being focused in center for zooming reasons.

Mouse position needs no further mucking with - it's passed in as related to the camera at the very beginning, so your
mouse x and mouse y coordinates now reflect mouse coordinates in game, not just on the screen.